URL scheme for viewing & creating note
