+++ Playground.playground/Contents.swift
@@ -1,0 +1,3 @@
+//: Playground - noun: a place where people can play
+
+import Cocoa
