+++ X2015.xcodeproj/xcuserdata/zhanghang.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -14,7 +14,7 @@
       <BreakpointProxy
          BreakpointExtensionID = "Xcode.Breakpoint.ExceptionBreakpoint">
          <BreakpointContent
-            shouldBeEnabled = "No"
+            shouldBeEnabled = "Yes"
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             scope = "0"
@@ -27,13 +27,13 @@
             shouldBeEnabled = "No"
             ignoreCount = "0"
             continueAfterRunningActions = "No"
-            filePath = "X2015/Views/UITableViewCell+Reusable.swift"
-            timestampString = "475609734.843141"
+            filePath = "TodayExtension/TodayTableViewController.swift"
+            timestampString = "476455089.280327"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "20"
-            endingLineNumber = "20"
-            landmarkName = "ReusableCell"
+            startingLineNumber = "99"
+            endingLineNumber = "99"
+            landmarkName = "TodayTableViewController"
             landmarkType = "3">
          </BreakpointContent>
       </BreakpointProxy>
@@ -43,13 +43,29 @@
             shouldBeEnabled = "No"
             ignoreCount = "0"
             continueAfterRunningActions = "No"
-            filePath = "X2015/ViewController/NoteTimelineTableViewController.swift"
-            timestampString = "476075204.302538"
+            filePath = "TodayExtension/TodayTableViewController.swift"
+            timestampString = "476455089.280327"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "173"
-            endingLineNumber = "173"
-            landmarkName = "updateNoteSelectionIfNeeded(_:deselectAll:)"
+            startingLineNumber = "85"
+            endingLineNumber = "85"
+            landmarkName = "TodayTableViewController"
+            landmarkType = "3">
+         </BreakpointContent>
+      </BreakpointProxy>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
+            shouldBeEnabled = "No"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "TodayExtension/TodayTableViewController.swift"
+            timestampString = "476455089.280327"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "84"
+            endingLineNumber = "84"
+            landmarkName = "widgetMarginInsetsForProposedMarginInsets(_:)"
             landmarkType = "5">
          </BreakpointContent>
       </BreakpointProxy>
@@ -59,77 +75,13 @@
             shouldBeEnabled = "No"
             ignoreCount = "0"
             continueAfterRunningActions = "No"
-            filePath = "X2015/ViewController/Theme/UIBarButtonItem+Theme.swift"
-            timestampString = "476078195.955164"
+            filePath = "X2015/AppDelegate/AppDelegate+Route.swift"
+            timestampString = "476454428.651166"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "12"
-            endingLineNumber = "12"
-            landmarkName = "configureThemeAppearance(_:)"
-            landmarkType = "5">
-            <Locations>
-               <Location
-                  shouldBeEnabled = "No"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "static ext.X2015.__ObjC.UIBarButtonItem.configureThemeAppearance (__ObjC.UIBarButtonItem.Type)(X2015.Theme) -&gt; ()"
-                  moduleName = "X2015"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/zhanghang/Workspace/X2015/X2015/ViewController/Theme/UIBarButtonItem+Theme.swift"
-                  timestampString = "476078202.867584"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "13"
-                  endingLineNumber = "13"
-                  offsetFromSymbolStart = "24">
-               </Location>
-               <Location
-                  shouldBeEnabled = "No"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "protocol witness for static X2015.AppearanceAdaptable.configureThemeAppearance &lt;A where A: X2015.AppearanceAdaptable&gt; (A.Type)(X2015.Theme) -&gt; () in conformance __ObjC.UIBarButtonItem : X2015.AppearanceAdaptable in X2015"
-                  moduleName = "X2015"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/zhanghang/Workspace/X2015/X2015/ViewController/Theme/UIBarButtonItem+Theme.swift"
-                  timestampString = "476078202.867844"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "12"
-                  endingLineNumber = "12"
-                  offsetFromSymbolStart = "0">
-               </Location>
-            </Locations>
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "X2015Kit/Utility/ManagedContextHelper.swift"
-            timestampString = "476104352.523246"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "49"
-            endingLineNumber = "49"
-            landmarkName = "createMainContext(_:)"
-            landmarkType = "5">
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            shouldBeEnabled = "Yes"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "X2015Kit/Utility/ManagedObjectContextBridge.swift"
-            timestampString = "476115793.628042"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "98"
-            endingLineNumber = "98"
-            landmarkName = "handleCFDidSaveContext()"
+            startingLineNumber = "17"
+            endingLineNumber = "17"
+            landmarkName = "createNote()"
             landmarkType = "5">
          </BreakpointContent>
       </BreakpointProxy>
