+++ X2015/AppDelegate/AppDelegate+CoreData.swift
@@ -1,0 +1,17 @@
+//
+//  AppDelegate+CoreData.swift
+//  X2015
+//
+//  Created by Hang Zhang on 2/6/16.
+//  Copyright © 2016 Zhang Hang. All rights reserved.
+//
+
+import X2015Kit
+
+extension AppDelegate {
+
+	func configureCoreData() {
+		Note.searchIndexEnabled = true
+	}
+
+}
